Renames left over ToastOverlay folder; Renames useCopyToast to useCopyWithFeedback for clarity.
