+++ cardstack/src/hooks/index.ts
@@ -4,7 +4,7 @@ export * from './use-lifetime-earnings-data';
 export * from './prepaid-card/useAuthToken';
 export * from './notifications-preferences/useUpdateNotificationPreferences';
 export * from './useMutationEffects';
-export * from './useCopyToast';
+export * from './useCopyWithFeedback';
 export * from './useBooleanState';
 export * from './useIsFetchingDataNewAccount';
 export * from './useGetAssetBalance';
